requirements.txt
requests==2.32.5
selenium==4.37.0
pytest==8.4.2
allure-pytest==2.35.1
